+++ {{cookiecutter.github_repository}}/{{cookiecutter.main_module}}/base/exceptions.py
@@ -1,8 +1,8 @@
 # Third Party Stuff
 from django.core.exceptions import PermissionDenied as DjangoPermissionDenied
 from django.http import Http404
-from django.utils.encoding import force_text
-from django.utils.translation import ugettext_lazy as _
+from django.utils.encoding import force_str
+from django.utils.translation import gettext_lazy as _
 from rest_framework import exceptions, status
 from rest_framework.response import Response
 
@@ -123,7 +123,7 @@ def format_exception(exc):
         for error_msg in exc.detail:
             detail["errors"].append({"message": error_msg})
     else:
-        detail["errors"].append({"message": force_text(exc.detail)})
+        detail["errors"].append({"message": force_str(exc.detail)})
 
     return detail