+++ {{cookiecutter.github_repository}}/{{cookiecutter.main_module}}/users/models.py
@@ -7,7 +7,7 @@ from django.contrib.auth.models import (
 from django.contrib.postgres.fields import CIEmailField
 from django.db import models
 from django.utils import timezone
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 # {{ cookiecutter.project_name }} Stuff
 from {{cookiecutter.main_module}}.base.models import UUIDModel